+++ GameRealisticMap.Studio/Modules/AssetBrowser/ViewModels/GdtDetailViewModel.cs
@@ -4,7 +4,6 @@ using System.IO;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
-using System.Windows.Media;
 using System.Windows.Media.Imaging;
 using System.Windows.Media.Media3D;
 using Caliburn.Micro;
@@ -14,13 +13,14 @@ using GameRealisticMap.Arma3.GameEngine.Materials;
 using GameRealisticMap.Studio.Modules.Arma3Data;
 using GameRealisticMap.Studio.Modules.Arma3Data.Services;
 using GameRealisticMap.Studio.Modules.AssetBrowser.Services;
+using GameRealisticMap.Studio.Modules.AssetBrowser.Services.Gdt;
 using GameRealisticMap.Studio.Modules.CompositionTool.ViewModels;
 using GameRealisticMap.Studio.Toolkit;
 using GameRealisticMap.Studio.UndoRedo;
+using GameRealisticMap.Toolkit;
 using Gemini.Framework;
 using Gemini.Framework.Services;
 using Gemini.Modules.UndoRedo;
-using HelixToolkit.Wpf;
 using Microsoft.Win32;
 using SixLabors.ImageSharp;
 using Color = System.Windows.Media.Color;
@@ -402,137 +402,19 @@ namespace GameRealisticMap.Studio.Modules.AssetBrowser.ViewModels
         {
             var tool = IoC.Get<PreviewToolViewModel>();
             var group = CreateBasePreview3d();
-            AddClutter(group);
+            Gdt3dPreviewHelper.AddClutter(group, ClutterList.Select(c => c.ToDefinition()));
             tool.Model3DGroup = group;
             IoC.Get<IShell>().ShowTool(tool);
             return Task.CompletedTask;
         }
-        private void AddClutter(Model3DGroup group)
-        {
-            var clutterPositions = Enumerable.Range(0, 20).Select(_ => new Point3D((Random.Shared.NextDouble() * 8) - 4, 0, (Random.Shared.NextDouble() * 8) - 4));
-
-            foreach (var clutter in ClutterList)
-            {
-                var count = (int)Math.Round(clutter.Probability * 20);
-                foreach (var pos in clutterPositions.Take(count))
-                {
-                    var preview = IoC.Get<IArma3Preview3D>().GetModel(clutter.Model.Path);
-                    if (preview != null)
-                    {
-                        var matrix = Matrix3D.Identity;
-                        var scale = clutter.ScaleMin + (Random.Shared.NextDouble() * (clutter.ScaleMax - clutter.ScaleMin));
-                        matrix.Scale(new Vector3D(scale, scale, scale));
-                        matrix.Translate(pos.ToVector3D());
-                        preview.Transform = new MatrixTransform3D(matrix);
-                        group.Children.Add(preview);
-                    }
-
-                }
-                clutterPositions = clutterPositions.Skip(count);
-            }
-        }
 
         private Model3DGroup CreateBasePreview3d()
         {
-            if (ImageColor == null || ImageNormal == null)
+            if (ImageColor == null)
             {
                 LoadImages();
             }
-
-            var group = new Model3DGroup();
-
-            var meshGdt = new MeshGeometry3D();
-            meshGdt.Positions = new Point3DCollection() {
-                new Point3D(0, 0, 0),
-                new Point3D(0, 0, 4),
-                new Point3D(4, 0, 4),
-                new Point3D(4, 0, 0),
-
-                new Point3D(-4, 0, -4),
-                new Point3D(-4, 0, 0),
-                new Point3D(0, 0, 0),
-                new Point3D(0, 0, -4),
-
-                new Point3D(0, 0, -4),
-                new Point3D(0, 0, 0),
-                new Point3D(4, 0, 0),
-                new Point3D(4, 0, -4),
-
-                new Point3D(-4, 0, 0),
-                new Point3D(-4, 0, 4),
-                new Point3D(0, 0, 4),
-                new Point3D(0, 0, 0),
-
-            };
-            meshGdt.TextureCoordinates = new PointCollection() {
-                new System.Windows.Point(0,0),
-                new System.Windows.Point(0,1),
-                new System.Windows.Point(1,1),
-                new System.Windows.Point(1,0),
-
-                new System.Windows.Point(0,0),
-                new System.Windows.Point(0,1),
-                new System.Windows.Point(1,1),
-                new System.Windows.Point(1,0),
-
-                new System.Windows.Point(0,0),
-                new System.Windows.Point(0,1),
-                new System.Windows.Point(1,1),
-                new System.Windows.Point(1,0),
-
-                new System.Windows.Point(0,0),
-                new System.Windows.Point(0,1),
-                new System.Windows.Point(1,1),
-                new System.Windows.Point(1,0)
-            };
-            meshGdt.TriangleIndices = new Int32Collection() {
-                0, 1, 2,
-                2, 3, 0,
-
-                4, 5, 6,
-                6, 7, 4,
-
-                8, 9, 10,
-                10, 11, 8,
-
-                12, 13, 14,
-                14, 15, 12
-            };
-
-            group.Children.Add(new GeometryModel3D(meshGdt, new DiffuseMaterial(new ImageBrush(ImageColor))));
-
-            var meshSat = new MeshGeometry3D();
-            meshSat.Positions = new Point3DCollection() {
-                new Point3D(4, 0, -4),
-                new Point3D(4, 0, 4),
-                new Point3D(6, 0, 4),
-                new Point3D(6, 0, -4),
-
-                new Point3D(-4, 0, 4),
-                new Point3D(-4, 0, 6),
-                new Point3D(6, 0, 6),
-                new Point3D(6, 0, 4),
-            };
-            meshSat.TextureCoordinates = new PointCollection() {
-                new System.Windows.Point(0,0),
-                new System.Windows.Point(0,1),
-                new System.Windows.Point(1,1),
-                new System.Windows.Point(1,0),
-
-                new System.Windows.Point(0,0),
-                new System.Windows.Point(0,1),
-                new System.Windows.Point(1,1),
-                new System.Windows.Point(1,0)
-            };
-            meshSat.TriangleIndices = new Int32Collection() {
-                0, 1, 2,
-                2, 3, 0,
-                4, 5, 6,
-                6, 7, 4,
-            };
-            group.Children.Add(new GeometryModel3D(meshSat, new DiffuseMaterial(new ImageBrush(FakeSatPreview))));
-
-            return group;
+            return Gdt3dPreviewHelper.CreateBasePreview3d(ImageColor!, FakeSatPreview ?? ImageColor!);
         }
 
         public Task BrowseImageColor()
