+++ GameRealisticMap/Toolkit/CaseConverter.cs
@@ -1,10 +1,6 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+﻿using System.Text;
 
-namespace GameRealisticMap.Studio.Toolkit
+namespace GameRealisticMap.Toolkit
 {
     public static class CaseConverter
     {
